Working build

+++ Assets/Scripts/UI/UI Toolkit/Controlls/InfoWindow/Controls/Building/FluidInfo.cs
@@ -15,8 +15,9 @@ public partial class FluidInfo : InfoWindowControl
 {
     List<FluidContainerElem> containers;
     int size = 100;
-    [UxmlAttribute] int Size 
-    { 
+    [UxmlAttribute]
+    int Size
+    {
         get => size;
         set
         {
@@ -27,13 +28,13 @@ public partial class FluidInfo : InfoWindowControl
                     q.container.style.width = size;
                     q.container.style.height = size;
                     q.filledMask.style.width = size;
-                    q.filledMask.style.height = size/2;
+                    q.filledMask.style.height = size / 2;
                     q.emptyLabel.style.width = size;
                     q.emptyLabel.style.height = size;
                     q.filledLabel.style.width = size;
                     q.filledLabel.style.height = size;
                 });
-        } 
+        }
     }
     public override void Open(object data)
     {
@@ -48,9 +49,9 @@ public partial class FluidInfo : InfoWindowControl
                 {
                     FluidType type = fluid.types[i];
                     containers.Add(CreateFluidIcon());
-                    containers[i].filledMask.style.color 
+                    containers[i].filledMask.style.color
                         = ToolkitUtils.resSkins.GetResourceColor(type);
-                    
+
                     var t = type;
                     var container = containers[i];
                     var x = i;
@@ -58,7 +59,7 @@ public partial class FluidInfo : InfoWindowControl
                     binding.sourceToUiConverters.AddConverter((ref Fluid flu) =>
                     {
                         Debug.Log((float)flu[t] / flu.capacities[x] * size);
-                        return new StyleLength((float)flu[t] / (float)flu.capacities[x] * size);
+                        return new StyleLength(flu[t] / (float)flu.capacities[x] * size);
                     });
                     SceneRefs.infoWindow.RegisterTempBinding(new BindingContext(container.filledMask, "style." + nameof(VisualElement.style.height)), binding, data);
 
@@ -76,25 +77,25 @@ public partial class FluidInfo : InfoWindowControl
 
     public FluidInfo()
     {
-        containers = new() { CreateFluidIcon(), CreateFluidIcon()};
+        containers = new() { CreateFluidIcon(), CreateFluidIcon() };
     }
 
     FluidContainerElem CreateFluidIcon()
     {
         Label emptyLabel, filledLabel;
-        VisualElement container = new() 
-        { 
-            name="container", 
-            style = 
-            { 
+        VisualElement container = new()
+        {
+            name = "container",
+            style =
+            {
                 width = size,
                 height = size
-            } 
+            }
         };
 
         VisualElement mask = new() { name = "empty-mask" };
-        mask.Add(emptyLabel= new Label("##") 
-        { 
+        mask.Add(emptyLabel = new Label("##")
+        {
             name = "empty-label",
             style =
             {
@@ -105,8 +106,8 @@ public partial class FluidInfo : InfoWindowControl
         container.Add(mask);
 
         mask = new() { name = "filled-mask" };
-        mask.Add(filledLabel = new Label("##") 
-        { 
+        mask.Add(filledLabel = new Label("##")
+        {
             name = "filled-label",
             style =
             {
@@ -116,7 +117,7 @@ public partial class FluidInfo : InfoWindowControl
         });
         container.Add(mask);
         Add(container);
-        return new FluidContainerElem() { container= container, filledMask=mask,emptyLabel= emptyLabel, filledLabel = filledLabel};
+        return new FluidContainerElem() { container = container, filledMask = mask, emptyLabel = emptyLabel, filledLabel = filledLabel };
     }
-   // void Create
+    // void Create
 }
